Links are not automatically pasted as markdown over valid urls (#189347)
* check if selected text is link
