+++ extensions/markdown-language-features/src/languageFeatures/copyFiles/copyPasteLinks.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import * as vscode from 'vscode';
-import { externalUriSchemes, createEditAddingLinksForUriList, getPasteUrlAsFormattedLinkSetting, PasteUrlAsFormattedLink } from './shared';
+import { createEditAddingLinksForUriList, getPasteUrlAsFormattedLinkSetting, PasteUrlAsFormattedLink, validateLink } from './shared';
 class PasteLinkEditProvider implements vscode.DocumentPasteEditProvider {
 
 	readonly id = 'insertMarkdownLink';
@@ -44,22 +44,6 @@ class PasteLinkEditProvider implements vscode.DocumentPasteEditProvider {
 		uriEdit.additionalEdit = pasteEdit.additionalEdits;
 		return uriEdit;
 	}
-}
-
-export function validateLink(urlList: string): { isValid: boolean; cleanedUrlList: string } {
-	let isValid = false;
-	let uri = undefined;
-	const trimmedUrlList = urlList?.trim(); //remove leading and trailing whitespace and new lines
-	try {
-		uri = vscode.Uri.parse(trimmedUrlList);
-	} catch (error) {
-		return { isValid: false, cleanedUrlList: urlList };
-	}
-	const splitUrlList = trimmedUrlList.split(' ').filter(item => item !== ''); //split on spaces and remove empty strings
-	if (uri) {
-		isValid = splitUrlList.length === 1 && !splitUrlList[0].includes('\n') && externalUriSchemes.includes(vscode.Uri.parse(splitUrlList[0]).scheme) && !!vscode.Uri.parse(splitUrlList[0]).authority;
-	}
-	return { isValid, cleanedUrlList: splitUrlList[0] };
 }
 
 export function registerLinkPasteSupport(selector: vscode.DocumentSelector,) {
