+++ extensions/markdown-language-features/src/test/markdownLink.test.ts
@@ -5,8 +5,7 @@
 import * as vscode from 'vscode';
 import * as assert from 'assert';
 import 'mocha';
-import { SkinnyTextDocument, checkSmartPaste, createEditAddingLinksForUriList, appendToLinkSnippet } from '../languageFeatures/copyFiles/shared';
-import { validateLink } from '../languageFeatures/copyFiles/copyPasteLinks';
+import { SkinnyTextDocument, checkSmartPaste, createEditAddingLinksForUriList, appendToLinkSnippet, validateLink } from '../languageFeatures/copyFiles/shared';
 suite('createEditAddingLinksForUriList', () => {
 
 	test('Markdown Link Pasting should occur for a valid link (end to end)', async () => {
@@ -168,6 +167,20 @@ suite('createEditAddingLinksForUriList', () => {
 			const range = new vscode.Range(0, 0, 0, 12);
 			const smartPaste = checkSmartPaste(skinnyDocument, range);
 			assert.strictEqual(smartPaste.pasteAsMarkdownLink, true);
+		});
+
+		test('Should evaluate pasteAsMarkdownLink as false for a valid selected link', () => {
+			skinnyDocument.getText = function () { return 'https://www.microsoft.com'; };
+			const range = new vscode.Range(0, 0, 0, 25);
+			const smartPaste = checkSmartPaste(skinnyDocument, range);
+			assert.strictEqual(smartPaste.pasteAsMarkdownLink, false);
+		});
+
+		test('Should evaluate pasteAsMarkdownLink as false for a valid selected link with trailing whitespace', () => {
+			skinnyDocument.getText = function () { return '   https://www.microsoft.com  '; };
+			const range = new vscode.Range(0, 0, 0, 30);
+			const smartPaste = checkSmartPaste(skinnyDocument, range);
+			assert.strictEqual(smartPaste.pasteAsMarkdownLink, false);
 		});
 
 		test('Should evaluate pasteAsMarkdownLink as false for no selection', () => {
